Refactor and answer to comments
