+++ research/conditional/moe_layers/cont_moe_designs/common_weighted_parameter_matrices.py
@@ -2,7 +2,7 @@ import dataclasses
 
 import torch
 from lizrd.core import misc, nn
-import lizrd.core.init
+import lizrd.core.initialization
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 from research.conditional.utils.misc_tools import stable_softmax_temperature
 
@@ -40,27 +40,27 @@ class ContinuousMoECommonWeightedParameters(ContinuousMoeBaseClass):
     def init_parameters(self):
         self.parameters_matrix_weight = nn.Parameter(torch.Tensor([0.5]))
         self.lin1 = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.dm
             )
         )
         self.lin2 = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.expert_size
             )
         )
         self.controller_base = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts), fan_in=self.dm * 2
             )
         )
         self.controller_merge = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts), fan_in=self.dm * 2
             )
         )
         self.controller_emit = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts), fan_in=self.dm * 2
             )
         )
