+++ research/conditional/moe_layers/cont_moe_designs/no_softmax_on_weights.py
@@ -1,7 +1,7 @@
 import torch
 
 from lizrd.core import misc, nn
-import lizrd.core.init
+import lizrd.core.initialization
 from lizrd.support import ash
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 
@@ -50,19 +50,21 @@ class ContinuousMoENosoftmax(ContinuousMoeBaseClass):
     def init_parameters(self):
         # lin1 is parameter, one dimension for experts of size dmodel to dff/n_experts
         self.lin1 = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.dm
             )
         )
 
         self.lin2 = nn.Parameter(
-            lizrd.core.init.get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.expert_size
             )
         )
         # controller: send a token of size dmodel to n_experts scalars
         self.controller = nn.Parameter(
-            lizrd.core.init.get_init_weight((self.dm, self.n_experts), fan_in=self.dm)
+            lizrd.core.initialization.get_init_weight(
+                (self.dm, self.n_experts), fan_in=self.dm
+            )
         )
         self.layernorm1 = nn.LayerNorm(self.dm)
         self.layernorm2 = nn.LayerNorm(self.dm)
