+++ research/conditional/moe_layers/cont_moe_designs/separate_merge_emit_weights.py
@@ -1,5 +1,5 @@
 from lizrd.core import nn, misc
-from lizrd.core.init import get_init_weight
+import lizrd.core.initialization
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 from research.conditional.utils.misc_tools import stable_softmax_temperature
 
@@ -11,18 +11,24 @@ class ContinuousMoEMergeDifferentlySimple(ContinuousMoeBaseClass):
 
     def init_parameters(self):
         self.lin1 = nn.Parameter(
-            get_init_weight((self.dm, self.n_experts, self.expert_size), fan_in=self.dm)
+            lizrd.core.initialization.get_init_weight(
+                (self.dm, self.n_experts, self.expert_size), fan_in=self.dm
+            )
         )
         self.lin2 = nn.Parameter(
-            get_init_weight(
+            lizrd.core.initialization.get_init_weight(
                 (self.dm, self.n_experts, self.expert_size), fan_in=self.expert_size
             )
         )
         self.controller_merge = nn.Parameter(
-            get_init_weight((self.dm, self.n_experts), fan_in=self.dm)
+            lizrd.core.initialization.get_init_weight(
+                (self.dm, self.n_experts), fan_in=self.dm
+            )
         )
         self.controller_emit = nn.Parameter(
-            get_init_weight((self.dm, self.n_experts), fan_in=self.dm)
+            lizrd.core.initialization.get_init_weight(
+                (self.dm, self.n_experts), fan_in=self.dm
+            )
         )
 
     def get_merge_and_emit_weights(self, x):
