+++ research/conditional/moe_layers/cont_moe_designs/separate_merge_emit_weights_common_base.py
@@ -1,5 +1,5 @@
 from lizrd.core import misc, nn
-from lizrd.core.init import get_init_weight
+from lizrd.core.initialization import get_init_weight
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 from research.conditional.utils.misc_tools import stable_softmax_temperature
