+++ research/conditional/moe_layers/cont_moe_designs/separate_weighted_parameter_matrices.py
@@ -2,7 +2,7 @@ import dataclasses
 
 import torch
 from lizrd.core import misc, nn
-from lizrd.core.init import get_init_weight
+from lizrd.core.initialization import get_init_weight
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 from research.conditional.utils.misc_tools import stable_softmax_temperature
