+++ research/conditional/train/configs/baselines/gpt/mini.yaml
@@ -1,8 +1,16 @@
 parent: research/conditional/train/configs/baselines/gpt/common.yaml
 md5_parent_hash: f10b3c37dd5f7aa8e7282a671a5b37e3
+cuda_visible: "2"
 
 params:
   dmodel: 256
   dff: 1024
   n_blocks: 4
   n_att_heads: 4
+  learning_rate: 4e-3
+  batch_size: 64
+  n_steps: 125_000
+  final_lr_fraction: 0.1
+  final_lr_step: 100_000
+  init_type: "truncated_normal"
+  init_scale: 1.0
